refactor(test_modeling_rule): move setup_rich_logging into logger module
