+++ demisto_sdk/commands/common/logger.py
@@ -1,7 +1,10 @@
 import logging
 import os
 import sys
+from pathlib import Path
 from typing import Optional
+
+from rich.logging import RichHandler
 
 DATE_FORMAT = '%Y-%m-%dT%H:%M:%SZ'
 
@@ -76,6 +79,44 @@ def logging_setup(
 
 
 logger: logging.Logger = logging_setup(verbose=1, quiet=False)
+
+
+def set_console_stream_handler(logger: logging.Logger, handler: RichHandler = RichHandler(rich_tracebacks=True)):
+    """Set the console stream handler.
+
+    Args:
+        logger (logging.Logger): Logger.
+        handler (RichHandler, optional): RichHandler. Defaults to RichHandler(rich_tracebacks=True).
+    """
+    console_handler_index = -1
+    for i, h in enumerate(logger.handlers):
+        if h.name == 'console-handler':
+            console_handler_index = i
+    if console_handler_index != -1:
+        logger.handlers[console_handler_index] = handler
+    else:
+        logger.addHandler(handler)
+
+
+def setup_rich_logging(verbosity: int, quiet: bool, log_path: Path, log_file_name: str):
+    """Override the default StreamHandler with the RichHandler.
+
+    Setup logging and then override the default StreamHandler with the RichHandler.
+
+    Args:
+        verbosity (int): The log level to output.
+        quiet (bool): If True, no logs will be output.
+        log_path (Path): Path to the directory where the log file will be created.
+        log_file_name (str): The filename of the log file.
+    """
+    logger = logging_setup(
+        verbose=verbosity,
+        quiet=quiet,
+        log_path=log_path,  # type: ignore[arg-type]
+        log_file_name=log_file_name
+    )
+    rich_handler = RichHandler(rich_tracebacks=True)
+    set_console_stream_handler(logger, rich_handler)
 
 
 # Python program to print
